Significant overhaul, adding memcache, RSS, and file uploads.
This update completes most of the Ackbar 1.0 work. It introduces
support for RSS, converting the post feed to XML. It also introduces
the AppEngine memcache API, which is currently configured to cache
basically all DataStore hits for as long as possible, and to automatically
re-cache everything as soon as you change something. Finally, this
commit introduces a new file upload feature (currently a bit hacky
because of API limitations) which uses the BlobStore API to serve
up arbitrary files.